refactor(calendar): share one calendar component between space and project
The project Calendar tab now renders the same CalendarView/WorkspaceCalendar
as the top-level /calendar page — the only difference is it passes projectIri
to filter to that project (and hides the project filter). Extracts the fetch/
filter/reschedule logic into CalendarView so both surfaces reuse it; removes
the old project-only TaskCalendar. Adds a ?project= filter to GET /calendar.

- api: CalendarController ?project= narrowing (+ CalendarTest case).
- pwa: new CalendarView container; calendar.tsx + project [id] Calendar tab
  both use it; delete TaskCalendar; drop the now-unused calendarDefinitions.
